move tests of store into separate test file
